+++ FabrieBank.DAL/Common/DTOs/DTOCurrencyRate.cs
@@ -3,9 +3,9 @@
     public class DTOCurrencyRate
     {
         public string? CurrencyCode { get; set; }
-        public double BanknoteBuyingRate { get; set; }
-        public double BanknoteSellingRate { get; set; }
+        public decimal BanknoteBuyingRate { get; set; }
+        public decimal BanknoteSellingRate { get; set; }
         public double ForexBuyingRate { get; set; }
-        public double ForexSellingRate { get; set; }
+        public decimal ForexSellingRate { get; set; }
     }
 }
